# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Milokhelo API
  version: "v1"
  description: >
    Milokhelo REST API v1 — sports match & tournament platform.
    Supports matches, tournaments, chat, venues (nearby search & booking), maps (mobile-submitted),
    calendar sync (device & third-party), notifications, invitations, feedback,
    teams, achievements, and hybrid OAuth (Google & Facebook) with backend token exchange & session cookie.
servers:
  - url: https://api.milokhelo.com/v1
    description: Production
  - url: http://localhost:4000/v1
    description: Local development
...
paths:
...
  /tournaments:
    post:
      tags: [Tournaments]
      summary: Create tournament
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentCreate"
      responses:
        "201":
          description: Tournament created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
    get:
      tags: [Tournaments]
      summary: List tournaments
      parameters:
        - name: sport
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema:
            type: string
            enum: [knockout, league]
      responses:
        "200":
          description: Tournaments array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournament"
  /tournaments/{id}:
    get:
      tags: [Tournaments]
      summary: Get tournament details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tournament
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
    patch:
      tags: [Tournaments]
      summary: Update tournament
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentUpdate"
      responses:
        "200":
          description: Tournament updated
    delete:
      tags: [Tournaments]
      summary: Cancel tournament
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Tournament cancelled
  /tournaments/{id}/register:
    post:
      tags: [Tournaments]
      summary: Register team/user
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TournamentRegister"
      responses:
        "200":
          description: Registered
  /tournaments/{id}/start:
    post:
      tags: [Tournaments]
      summary: Start tournament (generate brackets)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tournament started

  ################################################################
  # CHAT                                                           #
  ################################################################
  /chat/rooms:
    get:
      tags: [Chat]
      summary: List user's chat rooms
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Chat rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatRoom"
    post:
      tags: [Chat]
      summary: Create chat room
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRoomCreate"
      responses:
        "201":
          description: Chat room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatRoom"
  /chat/rooms/{roomId}/messages:
    get:
      tags: [Chat]
      summary: Fetch messages
      security:
        - cookieAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: before
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Messages array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      tags: [Chat]
      summary: Send message
      security:
        - cookieAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageCreate"
      responses:
        "201":
          description: Message saved
  /chat/messages/{messageId}:
    patch:
      tags: [Chat]
      summary: Edit message
      security:
        - cookieAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        "200":
          description: Message edited
    delete:
      tags: [Chat]
      summary: Delete (soft) message
      security:
        - cookieAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Message deleted

  ################################################################
  # VENUES (USER-FACING)                                           #
  ################################################################
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oauth2_google:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            profile: Access basic profile
            email: Access email address
    oauth2_facebook:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://www.facebook.com/v11.0/dialog/oauth
          tokenUrl: https://graph.facebook.com/v11.0/oauth/access_token
          scopes:
            public_profile: Access public profile
            email: Access email address
  schemas:
...
    TournamentCreate:
      type: object
      required: [title, sport, type, registrationWindow]
      properties:
        title: { type: string }
        sport: { type: string }
        sportCategory: { type: string }
        type:
          type: string
          enum: [knockout, league]
        rules: { type: string }
        registrationWindow:
          type: object
          required: [start, end]
          properties:
            start: { type: string, format: date-time }
            end: { type: string, format: date-time }
        entryFee: { type: number }
        prizePool: { type: string }
        minTeams: { type: integer }
        maxTeams: { type: integer }
        visibility:
          type: string
          enum: [public, private]
    TournamentUpdate:
      type: object
      properties:
        rules: { type: string }
        entryFee: { type: number }
        prizePool: { type: string }
        visibility:
          type: string
          enum: [public, private]
        status:
          type: string
          enum: [registration, ongoing, completed, cancelled]
        currentRound: { type: integer }
    TournamentRegister:
      type: object
      required: [teamId]
      properties:
        teamId: { type: string }
    Tournament:
      allOf:
        - $ref: "#/components/schemas/TournamentCreate"
        - type: object
          properties:
            id: { type: string }
            organizerId: { type: string }
            teams:
              type: array
              items: { type: string }
            matches:
              type: array
              items: { type: string }
            rounds: { type: integer }
            currentRound: { type: integer }
            bracket: { type: object }
            status:
              type: string
              enum: [registration, ongoing, completed, cancelled]
            chatRoomId: { type: string }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }

    #########################################################
    # CHAT SCHEMAS                                           #
    #########################################################
    ChatRoomCreate:
      type: object
      properties:
        name: { type: string }
        type:
          type: string
          enum: [match, tournament, team, dm, group]
        participants:
          type: array
          items: { type: string }
    ChatRoom:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type:
          type: string
          enum: [match, tournament, team, dm, group]
        participants:
          type: array
          items: { type: string }
        lastMessage: { type: string }
        mutedBy:
          type: array
          items: { type: string }
        isArchived: { type: boolean }
        presence:
          type: array
          items:
            type: object
            properties:
              userId: { type: string }
              status:
                type: string
                enum: [online, offline, typing]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ChatMessageCreate:
      type: object
      required: [content]
      properties:
        content: { type: string }
        messageType:
          type: string
          enum: [text, image, video, system]
        attachments:
          type: array
          items: { type: string }
        replyTo: { type: string }
    ChatMessage:
      allOf:
        - $ref: "#/components/schemas/ChatMessageCreate"
        - type: object
          properties:
            id: { type: string }
            roomId: { type: string }
            senderId: { type: string }
            reactions:
              type: array
              items:
                type: object
                properties:
                  userId: { type: string }
                  emoji: { type: string }
            deleted: { type: boolean }
            createdAt: { type: string, format: date-time }
            editedAt: { type: string, format: date-time }

    #########################################################
    # VENUE / MAP SCHEMAS                                    #
    #########################################################
...
security:
  - cookieAuth: []
...