# src/api/v1/modules/chat/index.js
/**
 * Chat Module - Comprehensive Stubs
 */
import { ChatRoomModel, ChatMessageModel } from '../shared/models.js';
import { asyncHandler } from '../../../../../common/utils/index.js';
import { HTTP_STATUS } from '../../../../../common/constants/index.js';
import express from 'express';

// Repository
export class ChatRepository {
  constructor(logger) {
    this.logger = logger.child({ context: 'ChatRepository' });
  }

  async createRoom(data) {
    const room = new ChatRoomModel(data);
    await room.save();
    return room.toObject();
  }

  async findRoomById(id) {
    return ChatRoomModel.findById(id).lean();
  }

  async findUserRooms(userId) {
    return ChatRoomModel.find({ participants: userId }).lean();
  }

  async createMessage(data) {
    const message = new ChatMessageModel(data);
    await message.save();
    return message.toObject();
  }

  async findMessages(roomId, limit = 50, before = null) {
    const query = { roomId, deleted: false };
    if (before) query.createdAt = { $lt: new Date(before) };
    return ChatMessageModel.find(query).sort({ createdAt: -1 }).limit(limit).lean();
  }

  async updateMessage(id, data) {
    return ChatMessageModel.findByIdAndUpdate(id, { ...data, editedAt: new Date() }, { new: true }).lean();
  }

  async deleteMessage(id) {
    return ChatMessageModel.findByIdAndUpdate(id, { deleted: true }, { new: true }).lean();
  }
}

// Service
export class ChatService {
  constructor(chatRepository, eventBus, logger) {
    this.chatRepository = chatRepository;
    this.eventBus = eventBus;
    this.logger = logger.child({ context: 'ChatService' });
  }

  async createRoom(data) {
    const room = await this.chatRepository.createRoom(data);
    await this.eventBus.publish('chat.room_created', { roomId: room._id });
    return room;
  }

  async getUserRooms(userId) {
    return this.chatRepository.findUserRooms(userId);
  }

  async sendMessage(roomId, senderId, data) {
    const message = await this.chatRepository.createMessage({ roomId, senderId, ...data });
    await this.eventBus.publish('chat.message_sent', { roomId, messageId: message._id, senderId });
    return message;
  }

  async getMessages(roomId, limit, before) {
    return this.chatRepository.findMessages(roomId, limit, before);
  }

  async editMessage(messageId, senderId, content) {
    const message = await this.chatRepository.updateMessage(messageId, { content });
    if (!message || message.senderId.toString() !== senderId) {
      throw new Error('Not authorized');
    }
    return message;
  }

  async deleteMessage(messageId, senderId) {
    const message = await ChatMessageModel.findById(messageId);
    if (!message || message.senderId.toString() !== senderId) {
      throw new Error('Not authorized');
    }
    return this.chatRepository.deleteMessage(messageId);
  }
}

// Controller
export class ChatController {
  constructor(chatService, logger) {
    this.chatService = chatService;
    this.logger = logger.child({ context: 'ChatController' });
  }

  getRooms() {
    return asyncHandler(async (req, res) => {
      const userId = req.session?.userId;
      const rooms = await this.chatService.getUserRooms(userId);
      res.status(HTTP_STATUS.OK).json(rooms);
    });
  }

  createRoom() {
    return asyncHandler(async (req, res) => {
      const room = await this.chatService.createRoom(req.body);
      res.status(HTTP_STATUS.CREATED).json(room);
    });
  }

  getMessages() {
    return asyncHandler(async (req, res) => {
      const { roomId } = req.params;
      const { limit, before } = req.query;
      const messages = await this.chatService.getMessages(roomId, limit, before);
      res.status(HTTP_STATUS.OK).json(messages);
    });
  }

  sendMessage() {
    return asyncHandler(async (req, res) => {
      const userId = req.session?.userId;
      const { roomId } = req.params;
      const message = await this.chatService.sendMessage(roomId, userId, req.body);
      res.status(HTTP_STATUS.CREATED).json(message);
    });
  }

  editMessage() {
    return asyncHandler(async (req, res) => {
      const userId = req.session?.userId;
      const { messageId } = req.params;
      const message = await this.chatService.editMessage(messageId, userId, req.body.content);
      res.status(HTTP_STATUS.OK).json(message);
    });
  }

  deleteMessage() {
    return asyncHandler(async (req, res) => {
      const userId = req.session?.userId;
      const { messageId } = req.params;
      await this.chatService.deleteMessage(messageId, userId);
      res.status(HTTP_STATUS.NO_CONTENT).send();
    });
  }
}

// Routes
export function createChatRoutes(controller) {
  const router = express.Router();
  router.get('/rooms', controller.getRooms());
  router.post('/rooms', controller.createRoom());
  router.get('/rooms/:roomId/messages', controller.getMessages());
  router.post('/rooms/:roomId/messages', controller.sendMessage());
  router.patch('/messages/:messageId', controller.editMessage());
  router.delete('/messages/:messageId', controller.deleteMessage());
  return router;
}

// Module Initializer
export function initializeChatModule(container) {
  const logger = container.resolve('logger');
  const eventBus = container.resolve('eventBus');

  container.registerSingleton('chatRepository', () => new ChatRepository(logger));
  container.registerSingleton('chatService', () => {
    const repo = container.resolve('chatRepository');
    return new ChatService(repo, eventBus, logger);
  });
  container.registerSingleton('chatController', () => {
    const service = container.resolve('chatService');
    return new ChatController(service, logger);
  });

  logger.info('Chat module initialized');
}